Created core directory for :network:library and moved :mvvm:library into it too
