+++ application/app/build.gradle.kts
@@ -2,7 +2,7 @@ androidBinary(
     packageName = "com.stepango.blockme",
     projectDependencies = deps(
         project(":root-library"),
-        project(":mvvm:library"),
+        project(":core:mvvm:library"),
         project(":common:extensions:android-util")
     )
 )
