+++ application/core/mvvm/library/.gitignore
