+++ application/core/mvvm/library/proguard-rules.pro
