+++ application/core/mvvm/library/src/main/java/com/stepango/blockme/core/mvvm/library/livedata/SingleLiveData.kt
@@ -1,0 +1,74 @@
+/*
+ * Copyright 2019 vmadalin.com
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.stepango.blockme.core.mvvm.library.livedata
+
+import androidx.annotation.MainThread
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Observer
+import java.util.concurrent.atomic.AtomicBoolean
+import timber.log.Timber
+
+/**
+ * A lifecycle-aware observable that sends only new updates after subscription, used for events
+ * like navigation and Snackbar messages.
+ *
+ * This avoids a common problem with events: on configuration change (like rotation) an update
+ * can be emitted if the observer is active. This LiveData only calls the observable if there's an
+ * explicit call to setValue() or call().
+ *
+ * Note that only one observer is going to be notified of changes.
+ *
+ * @see MutableLiveData
+ */
+class SingleLiveData<T> : MutableLiveData<T>() {
+
+    private val pending = AtomicBoolean()
+
+    /**
+     * Adds the given observer to the observers list within the lifespan of the given
+     * owner. The events are dispatched on the main thread. If LiveData already has data
+     * set, it will be delivered to the observer.
+     *
+     * @param owner The LifecycleOwner which controls the observer
+     * @param observer The observer that will receive the events
+     * @see MutableLiveData.observe
+     */
+    @MainThread
+    override fun observe(owner: LifecycleOwner, observer: Observer<in T>) {
+        if (hasActiveObservers()) {
+            Timber.w("Multiple observers registered but only one will be notified of changes.")
+        }
+        super.observe(owner, { t ->
+            if (pending.compareAndSet(true, false)) {
+                observer.onChanged(t)
+            }
+        })
+    }
+
+    /**
+     * Sets the value. If there are active observers, the value will be dispatched to them.
+     *
+     * @param value The new value
+     * @see MutableLiveData.setValue
+     */
+    @MainThread
+    override fun setValue(value: T?) {
+        pending.set(true)
+        super.setValue(value)
+    }
+}
