+++ application/core/mvvm/library/src/main/java/com/stepango/blockme/core/mvvm/library/ui/BaseViewHolder.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2019 vmadalin.com
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.stepango.blockme.core.mvvm.library.ui
+
+import androidx.databinding.ViewDataBinding
+import androidx.recyclerview.widget.RecyclerView
+
+/**
+ * Base view holder to standardize and simplify initialization for this component.
+ *
+ * @param binding View data binding generated class instance.
+ * @see RecyclerView.ViewHolder
+ */
+abstract class BaseViewHolder<T : ViewDataBinding>(
+    val binding: T
+) : RecyclerView.ViewHolder(binding.root)
