+++ application/core/network/library/.gitignore
