+++ application/core/network/library/build.gradle.kts
@@ -1,0 +1,5 @@
+library(
+    projectDependencies = deps(
+        project(":common:util")
+    )
+)
