+++ application/core/network/library/src/main/java/com/stepango/blockme/core/network/library/response/BaseResponse.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2019 vmadalin.com
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.stepango.blockme.core.network.library.response
+
+import com.stepango.blockme.common.util.annotations.OpenForTesting
+
+/**
+ * Generic network response for any type data [T].
+ *
+ * @param code The HTTP status code of the returned result.
+ * @param status A string description of the call status.
+ * @param message A more descriptive message of the failure call status.
+ * @param data The results returned by the call.
+ */
+@OpenForTesting
+data class BaseResponse<T>(
+    val code: Any,
+    val status: String,
+    val message: String,
+    val data: DataResponse<T>
+)
