--- application/demo-library/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.stepango.mylibrary">
-
-</manifest>
