--- application/mvvm/library/src/main/java/com/stepango/blockme/mvvm/library/ui/BaseFragment.kt
@@ -1,124 +1,0 @@
-/*
- * Copyright 2019 vmadalin.com
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.stepango.blockme.mvvm.library.ui
-
-import android.content.Context
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.annotation.LayoutRes
-import androidx.appcompat.app.AppCompatActivity
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModel
-import javax.inject.Inject
-
-/**
- * Base fragment to standardize and simplify initialization for this component.
- *
- * @param layoutId Layout resource reference identifier.
- * @see Fragment
- */
-abstract class BaseFragment<B : ViewDataBinding, M : ViewModel>(
-    @LayoutRes
-    private val layoutId: Int
-) : Fragment() {
-
-    @Inject
-    lateinit var viewModel: M
-    lateinit var viewBinding: B
-
-    /**
-     * Called to initialize dagger injection dependency graph when fragment is attached.
-     */
-    abstract fun onInitDependencyInjection()
-
-    /**
-     * Called to Initialize view data binding variables when fragment view is created.
-     */
-    abstract fun onInitDataBinding()
-
-    /**
-     * Called to have the fragment instantiate its user interface view.
-     *
-     * @param inflater The LayoutInflater object that can be used to inflate any views in the fragment.
-     * @param container If non-null, this is the parent view that the fragment's UI should be
-     * attached to. The fragment should not add the view itself, but this can be used to generate
-     * the LayoutParams of the view.
-     * @param savedInstanceState If non-null, this fragment is being re-constructed from a previous
-     * saved state as given here.
-     *
-     * @return Return the View for the fragment's UI, or null.
-     *
-     * @see Fragment.onCreateView
-     */
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        viewBinding = DataBindingUtil.inflate(inflater, layoutId, container, false)
-        viewBinding.lifecycleOwner = viewLifecycleOwner
-        return viewBinding.root
-    }
-
-    /**
-     * Called when a fragment is first attached to its context.
-     *
-     * @param context The application context.
-     *
-     * @see Fragment.onAttach
-     */
-    override fun onAttach(context: Context) {
-        super.onAttach(context)
-        onInitDependencyInjection()
-    }
-
-    /**
-     * Called to have the fragment instantiate its user interface view.
-     *
-     * @param view The view returned by onCreateView(LayoutInflater, ViewGroup, Bundle)}.
-     * @param savedInstanceState If non-null, this fragment is being re-constructed
-     * from a previous saved state as given here.
-     * @see Fragment.onViewCreated
-     */
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        onInitDataBinding()
-    }
-
-    /**
-     * Return the [AppCompatActivity] this fragment is currently associated with.
-     *
-     * @throws IllegalStateException if not currently associated with an activity or if associated
-     * only with a context.
-     * @throws TypeCastException if the currently associated activity don't extend [AppCompatActivity].
-     *
-     * @see requireActivity
-     */
-    fun requireCompatActivity(): AppCompatActivity {
-        requireActivity()
-        val activity = requireActivity()
-        if (activity is AppCompatActivity) {
-            return activity
-        } else {
-            throw TypeCastException("Main activity should extend from 'AppCompatActivity'")
-        }
-    }
-}
